Create output statement and then optimitize it
The example file shows the variables and the output statement that needs to be created. The variables should then be created and the output statement should also be constructed to match the output necessary. Then, to optimize the output statement, the firstName and lastname variables will be combined to a fullName variable and the age should be calculated using the thisYear and birthYear variables. The output statement should then be properly updated in order to use these new variables. The spec file's tests should all then be completely enabled in order to ensure that the output is correct for both output statements.